requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.4
djangorestframework==3.15.1
dnspython==2.6.1
drf-yasg==1.21.7
email-validator==2.0.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.1.1
oauthlib==3.2.2
openai==0.27.4
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
SQLAlchemy==2.0.15
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.9.4
